requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
python_dateutil==2.8.2
requests==2.27.1
scikit_learn==1.1.2
seaborn==0.11.2
streamlit==1.12.2
tensorflow==2.10.0
